backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pytest==7.4.2 
validators==0.22.0 
mysql-connector-python==8.1.0 
datetime==5.2.0
python-dotenv==1.0.0 
